make new file dummydata and refactor
