+++ lib/core/models/models.dart
@@ -4,6 +4,7 @@ export 'activity_overview.dart';
 export 'church.dart';
 export 'column.dart';
 export 'date_time_converter.dart';
+export 'dummy_data.dart';
 export 'location.dart';
 export 'membership.dart';
 export 'song_details_data.dart';
